feat(mongodb): multi-tab panels for collections
- MongoDBCollectionView: extract single-collection right panel (query /
  indexes / document editor / pagination / AI filter) as its own component
  so each tab keeps independent state
- MongoDBTabContent: tabs array + closable tab bar, one tab per collection
  (or per database query); v-show keeps each collection's query/pagination
  state alive; drop database/collection closes related tabs
- reuses db.tabClose i18n key from #543; no SQL import/export (Mongo scope)
